+++ storage_drivers/ontap/ontap_nas_qtree.go
@@ -1580,10 +1580,6 @@ func (d *NASQtreeStorageDriver) GetUpdateType(_ context.Context, driverOrig stor
 		return bitmap
 	}
 
-	if d.Config.DataLIF != dOrig.Config.DataLIF {
-		bitmap.Add(storage.VolumeAccessInfoChange)
-	}
-
 	if d.Config.Password != dOrig.Config.Password {
 		bitmap.Add(storage.PasswordChange)
 	}